requirements.txt
torch  
torchvision  
numpy  
pandas  
plotly  
umap-learn  
faiss-cpu  
streamlit  
open-clip-torch  
datumaro
transformers
